Rename models.py to doctypes.py
